requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.dev.in
#
asgiref==3.7.2
    # via django
black==23.7.0
    # via -r requirements.dev.in
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via black
django==4.2.4
    # via
    #   -r requirements.in
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   easy-thumbnails
django-stubs==4.2.3
    # via
    #   -r requirements.dev.in
    #   djangorestframework-stubs
django-stubs-ext==4.2.2
    # via django-stubs
djangorestframework==3.14.0
    # via -r requirements.in
djangorestframework-stubs==3.14.2
    # via -r requirements.dev.in
easy-thumbnails==2.8.5
    # via -r requirements.in
factory-boy==3.3.0
    # via -r requirements.dev.in
faker==19.3.0
    # via factory-boy
flake8==6.1.0
    # via -r requirements.dev.in
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements.dev.in
mccabe==0.7.0
    # via flake8
mypy==1.5.1
    # via
    #   -r requirements.dev.in
    #   django-stubs
    #   djangorestframework-stubs
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.1
    # via
    #   black
    #   pytest
pathspec==0.11.2
    # via black
pillow==10.0.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
platformdirs==3.10.0
    # via black
pluggy==1.2.0
    # via pytest
psycopg2==2.9.7
    # via -r requirements.in
pycodestyle==2.11.0
    # via flake8
pyflakes==3.1.0
    # via flake8
pytest==7.4.0
    # via
    #   -r requirements.dev.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.dev.in
python-dateutil==2.8.2
    # via faker
pytz==2023.3
    # via djangorestframework
requests==2.31.0
    # via djangorestframework-stubs
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
types-pytz==2023.3.0.1
    # via django-stubs
types-pyyaml==6.0.12.11
    # via
    #   django-stubs
    #   djangorestframework-stubs
types-requests==2.31.0.2
    # via djangorestframework-stubs
types-urllib3==1.26.25.14
    # via types-requests
typing-extensions==4.7.1
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework-stubs
    #   mypy
urllib3==2.0.4
    # via requests
